# docs/swagger.yaml
basePath: /
definitions:
...
  dto.ChangePasswordRequest:
    properties:
      old_password:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - old_password
    - password
    type: object
...
  dto.ProfileResponse:
    properties:
      full_name:
        type: string
      phone:
        type: string
      photo:
        type: string
    type: object
...
  dto.Response:
    properties:
      data: {}
      error:
        type: string
      isSuccess:
        type: boolean
      message:
        type: string
    type: object
  dto.SetPinRequest:
    properties:
      old_pin:
        type: string
      pin_hash:
        type: string
    required:
    - pin_hash
    type: object
  dto.UserInfoResponse:
    properties:
      current_balance:
        type: integer
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      phone:
        type: string
      photo:
        type: string
      pin_hash:
        type: string
      wallet_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Backend Vanwallet  using Gin
  license:
    name: MIT
  title: Vanwallet
  version: "1.0"
paths:
...
  /profile:
    get:
      consumes:
      - application/json
      description: Retrieves current authentication details' full name, telephone
        connection code, and user avatar endpoint.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.ProfileResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get user profile details
      tags:
      - Profile
  /profile/change/pin:
    patch:
      consumes:
      - application/json
      description: 'First-time setup: provide only pin_hash. Changing existing PIN:
        provide old_pin + pin_hash.'
      parameters:
      - description: PIN update payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SetPinRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Update user secondary authorization pin
      tags:
      - Profile
  /profile/edit:
    patch:
      consumes:
      - multipart/form-data
      description: Updates system details including full name text, phone data info,
        and multipart image form attachments.
      parameters:
      - description: Updated full identity name representation
        in: formData
        name: full_name
        type: string
      - description: Target telecommunications contact identity sequence
        in: formData
        name: phone
        type: string
      - description: Binary source file image attachment content
        in: formData
        name: photo
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      - BearerAuth: []
      summary: Modify active user profile records
      tags:
      - Profile
  /profile/info:
    get:
      consumes:
      - application/json
      description: Assembles structured systemic components including identities,
        security states, and financial metrics.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserInfoResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Get unified system user statistics context
      tags:
      - Profile
  /profile/password:
    patch:
      consumes:
      - application/json
      description: Modifies internal account validation strings. Requires old confirmation
        verification strings.
      parameters:
      - description: Password structure swap payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.Response'
      security:
      - BearerAuth: []
      summary: Modify security entry password credentials
      tags:
      - Profile
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and your JWT. Example: "Bearer
      eyJhbGci..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"